requirements.txt
blinker==1.7.0
bson==0.5.10
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
Flask==3.0.0
fonttools==4.47.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
Werkzeug==3.0.1
zipp==3.17.0
